Properties can now also be called 'cleanly' from moonscript.
Previously, in moonscript properties would have to be defined as follows
in order not to clash with the DSL-like Lua-syntax:

property('description')
    generators: { ... }
    check: check_function

or:

property'description'
    generators: { ... }
    check: check_function

Neither solution was nice to use and/or confusing. This commit fixes that by
allowing a property to also be defined without the DSL syntax.

Now you can write:

property 'description'
    generators: { ... }
    check: check_function